Move key stuff to its own file
